New module for build_demand. removes demand building from add_electricity.
